# src/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.ChannelPayload:
    properties:
      all_services:
        type: boolean
      chat_id:
        type: string
      events:
        items:
          type: string
        type: array
      headers:
        additionalProperties:
          type: string
        type: object
      method:
        type: string
      name:
        type: string
      services:
        items:
          type: string
        type: array
      sub_chat_id:
        type: string
      token:
        type: string
      type:
        $ref: '#/definitions/notify.ChannelType'
      url:
        type: string
    type: object
  handler.PatchEnabledPayload:
    properties:
      enabled:
        type: boolean
    type: object
...
  handler.TestResponse:
    properties:
      detail:
        type: string
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  handler.messageJSON:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  notify.ChannelType:
    enum:
    - slack
    - discord
    - telegram
    - custom
    type: string
    x-enum-varnames:
    - ChannelTypeSlack
    - ChannelTypeDiscord
    - ChannelTypeTelegram
    - ChannelTypeCustom
  notify.HistoryEntry:
    properties:
      channel_id:
        type: string
      channel_name:
        type: string
      channel_type:
        type: string
      duration_ms:
        description: 含 retry 的總耗時
        type: integer
      error:
        description: 失敗原因
        type: string
      event:
        description: started/stopped/failed/restarted/test
        type: string
      service:
        description: nginx.service
        type: string
      status:
        description: success/failure
        type: string
      timestamp:
        description: RFC3339 UTC
        type: string
    type: object
  notify.HistoryResult:
    properties:
      data:
        items:
          $ref: '#/definitions/notify.HistoryEntry'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
info:
  contact: {}
  title: Linux Service Manager API
  version: 1.0.0
paths:
...
  /notify/channels:
    get:
      description: 列出所有通知 Channel（Telegram token 遮罩顯示）。`read` scope Token 可用。
      produces:
      - application/json
      responses:
        "200":
          description: '{\"data\": [Channel...]}'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未驗證
          schema:
            $ref: '#/definitions/handler.messageJSON'
        "500":
          description: notify 未設定
          schema:
            $ref: '#/definitions/handler.messageJSON'
      security:
      - BearerAuth: []
      summary: 列出通知 Channels
      tags:
      - Notifications
    post:
      consumes:
      - application/json
      description: 建立新的通知 Channel（Slack/Discord/Telegram/Custom webhook）。需 `full`
        scope Token。
      parameters:
      - description: Channel 設定
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.ChannelPayload'
      produces:
      - application/json
      responses:
        "201":
          description: '{\"data\": Channel}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 驗證失敗
          schema:
            $ref: '#/definitions/handler.messageJSON'
        "401":
          description: 未驗證
          schema:
            $ref: '#/definitions/handler.messageJSON'
        "403":
          description: 唯讀 Token 權限不足
          schema:
            $ref: '#/definitions/handler.messageJSON'
        "409":
          description: 已達 Channel 上限（20）
          schema:
            $ref: '#/definitions/handler.messageJSON'
        "500":
          description: 建立失敗
          schema:
            $ref: '#/definitions/handler.messageJSON'
      security:
      - BearerAuth: []
      summary: 建立通知 Channel
      tags:
      - Notifications
  /notify/channels/{id}:
    delete:
      description: 刪除 Channel（歷史發送紀錄保留）。需 `full` scope Token。
      parameters:
      - description: Channel ID（UUID）
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.messageJSON'
        "401":
          description: 未驗證
          schema:
            $ref: '#/definitions/handler.messageJSON'
        "403":
          description: 唯讀 Token 權限不足
          schema:
            $ref: '#/definitions/handler.messageJSON'
        "404":
          description: Channel 不存在
          schema:
            $ref: '#/definitions/handler.messageJSON'
        "500":
          description: 刪除失敗
          schema:
            $ref: '#/definitions/handler.messageJSON'
      security:
      - BearerAuth: []
      summary: 刪除通知 Channel
      tags:
      - Notifications
    patch:
      consumes:
      - application/json
      description: 切換 Channel 的 enabled 狀態。需 `full` scope Token。
      parameters:
      - description: Channel ID（UUID）
        in: path
        name: id
        required: true
        type: string
      - description: '{\'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.PatchEnabledPayload'
      produces:
      - application/json
      responses:
        "200":
          description: '{\"data\": Channel}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 請求無效
          schema:
            $ref: '#/definitions/handler.messageJSON'
        "401":
          description: 未驗證
          schema:
            $ref: '#/definitions/handler.messageJSON'
        "403":
          description: 唯讀 Token 權限不足
          schema:
            $ref: '#/definitions/handler.messageJSON'
        "404":
          description: Channel 不存在
          schema:
            $ref: '#/definitions/handler.messageJSON'
        "500":
          description: 更新失敗
          schema:
            $ref: '#/definitions/handler.messageJSON'
      security:
      - BearerAuth: []
      summary: 啟用/停用通知 Channel
      tags:
      - Notifications
    put:
      consumes:
      - application/json
      description: 覆寫更新 Channel 設定。Telegram token 留空或以 `****` 開頭時保留原值。需 `full` scope
        Token。
      parameters:
      - description: Channel ID（UUID）
        in: path
        name: id
        required: true
        type: string
      - description: 新的 Channel 設定
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.ChannelPayload'
      produces:
      - application/json
      responses:
        "200":
          description: '{\"data\": Channel}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 驗證失敗
          schema:
            $ref: '#/definitions/handler.messageJSON'
        "401":
          description: 未驗證
          schema:
            $ref: '#/definitions/handler.messageJSON'
        "403":
          description: 唯讀 Token 權限不足
          schema:
            $ref: '#/definitions/handler.messageJSON'
        "404":
          description: Channel 不存在
          schema:
            $ref: '#/definitions/handler.messageJSON'
        "500":
          description: 更新失敗
          schema:
            $ref: '#/definitions/handler.messageJSON'
      security:
      - BearerAuth: []
      summary: 更新通知 Channel
      tags:
      - Notifications
  /notify/channels/{id}/test:
    post:
      description: 向指定 Channel 發送測試通知（不寫入 history）。需 `full` scope Token。\n\n**注意**：發送失敗回
        **502 Bad Gateway**（非 500）。
      parameters:
      - description: Channel ID（UUID）
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.TestResponse'
        "401":
          description: 未驗證
          schema:
            $ref: '#/definitions/handler.messageJSON'
        "403":
          description: 唯讀 Token 權限不足
          schema:
            $ref: '#/definitions/handler.messageJSON'
        "404":
          description: Channel 不存在
          schema:
            $ref: '#/definitions/handler.messageJSON'
        "502":
          description: 發送失敗（success=false, error=原因）
          schema:
            $ref: '#/definitions/handler.TestResponse'
      security:
      - BearerAuth: []
      summary: 測試通知 Channel
      tags:
      - Notifications
  /notify/history:
    get:
      description: 分頁查詢通知發送紀錄，可依 Channel 與結果篩選。`read` scope Token 可用。
      parameters:
      - default: 1
        description: 頁碼（預設 1）
        in: query
        name: page
        type: integer
      - default: 30
        description: 每頁筆數（預設 30，上限 100）
        in: query
        name: limit
        type: integer
      - description: 依 Channel ID 篩選
        in: query
        name: channel_id
        type: string
      - description: all|success|failure（預設 all）
        enum:
        - all
        - success
        - failure
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notify.HistoryResult'
        "400":
          description: 參數無效
          schema:
            $ref: '#/definitions/handler.messageJSON'
        "401":
          description: 未驗證
          schema:
            $ref: '#/definitions/handler.messageJSON'
        "500":
          description: 查詢失敗
          schema:
            $ref: '#/definitions/handler.messageJSON'
      security:
      - BearerAuth: []
      summary: 查詢通知發送紀錄
      tags:
      - Notifications
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"